Make all unused palette colors be green
If the BG ram is bad, or if the BG ram is good but the path from it to the pal chips is fautly, it can cause the wrong palette/palette color to be used for the background.  This fixes an issue where you could end up with a white background color, which makes it impossible to see any text on screen.  Instead the fault will cause the background to be green indicating something is wrong.