Implement error addresses
After playing the error code, jump to a specific address for the
given error, which will in turn loop to itself.  Address lines can
then be probed with a logic probe to determine what address the
diag rom stopped on.  The stop address can then be looked up in
docs/error_codes.md to determine the code error.  This is specifically
useful if both the screen and sound are not in a working state.