Add MCU tests to ddragon2
I had to exclude the checksum test as dd2's mcu doesn't provide
the data.

The weird thing is the mcu on dd2 has code to calculate a checksum
of the rom and store it at the first byte of the com ram, just like
dd1's mcu did.  However nothing ever seems to call the code so it
never happens.